refactor: add custom hook for page scroll tracking, refactor header vars
