+++ src/app/[locale]/login/page.tsx
@@ -1,6 +1,5 @@
 import AuthForm from '@/components/auth-form/auth-form';
 
-
 const LoginForm = () => {
   return <AuthForm type="login" />;
 };
