+++ src/app/[locale]/registration/page.tsx
@@ -1,6 +1,5 @@
 import AuthForm from '@/components/auth-form/auth-form';
 
-
 const RegistrationForm = () => {
   return <AuthForm type="register" />;
 };
