--- src/components/example/auth-form/auth-form.tsx
@@ -1,84 +1,0 @@
-'use client';
-
-import type { FC } from 'react';
-import { type FieldErrors, useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { TextField, Button, Box, Typography, Link } from '@mui/material';
-import { type SignInInput, type SignUpInput, signInSchema, signUpSchema } from '@/schema/auth-schema';
-
-interface IProps {
-  type: 'login' | 'register';
-}
-
-type TFormValues = SignInInput | SignUpInput;
-
-const AuthForm: FC<IProps> = ({ type }) => {
-  const isLogin = type === 'login';
-  const schema = isLogin ? signInSchema : signUpSchema;
-
-  const {
-    register,
-    formState: { errors, isValid },
-  } = useForm<TFormValues>({
-    resolver: zodResolver(schema),
-    mode: 'onChange',
-  });
-
-  return (
-    <Box
-      component="form"
-      sx={{ width: 400, mx: 'auto', mt: 8 }}
-      onSubmit={(e) => {
-        e.preventDefault();
-      }}
-    >
-      <Typography variant="h5" align="center" mb={2}>
-        {isLogin ? 'Sign In' : 'Sign Up'}
-      </Typography>
-
-      {!isLogin && (
-        <TextField
-          label="Name"
-          fullWidth
-          margin="normal"
-          {...register('name')}
-          helperText={(errors as FieldErrors<SignUpInput>).name?.message}
-          error={!!(errors as FieldErrors<SignUpInput>).name}
-        />
-      )}
-
-      <TextField
-        label="Email"
-        fullWidth
-        margin="normal"
-        {...register('email')}
-        helperText={errors.email?.message}
-        error={!!errors.email}
-      />
-
-      <TextField
-        label="Password"
-        type="password"
-        fullWidth
-        margin="normal"
-        {...register('password')}
-        helperText={errors.password?.message}
-        error={!!errors.password}
-      />
-
-      <Box mt={2} display="flex" justifyContent="center">
-        <Button type="submit" variant="contained" disabled={!isValid} fullWidth>
-          {isLogin ? 'Sign In' : 'Sign Up'}
-        </Button>
-      </Box>
-
-      <Box sx={{ textAlign: 'center' }}>
-        <Link href="#" variant="body1" sx={{ display: 'block', mt: 2 }}>
-          {isLogin ? "Don't have an account? Sign Up" : 'Already have an account? Sign In'}
-        </Link>
-      </Box>
-    </Box>
-  );
-};
-
-export default AuthForm;
