--- src/hooks/example-hook.ts
@@ -1,1 +1,0 @@
-// custom hook will be here
